# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Toutci API v1",
    "version": "1.1.0",
    "description": "Contrat fonctionnel de référence pour les clients web et mobiles Toutci."
  },
  "servers": [
    {
      "url": "https://restauci.vercel.app/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "partnerLogin",
        "tags": [
          "Auth partenaire"
        ],
        "summary": "Connecter un partenaire ou administrateur",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Session absente ou expirée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requêtes dépassée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "partnerLogout",
        "tags": [
          "Auth partenaire"
        ],
        "summary": "Déconnecter le compte partenaire",
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Session absente ou expirée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requêtes dépassée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/livreur/auth/logout": {
      "post": {
        "operationId": "logoutDriver",
        "tags": [
          "Auth livreur"
        ],
        "summary": "Révoquer une session livreur",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Session absente ou expirée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requêtes dépassée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiSuccess": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {},
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "required": [
          "success",
          "error",
          "code"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": [
          "total",
          "page",
          "limit",
          "totalPages",
          "hasNext",
          "hasPrev"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrev": {
            "type": "boolean"
          }
        }
      },
...
      "TokenTransport": {
        "type": "string",
        "enum": [
          "cookie",
          "json"
        ],
        "default": "cookie"
      },
      "PartnerLoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "rememberMe": {
            "type": "boolean"
          },
          "otp": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "description": "Obligatoire pour les administrateurs."
          }
        }
      },
...
      "DriverRefreshRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tokenTransport": {
            "$ref": "#/components/schemas/TokenTransport"
          },
          "refreshToken": {
            "type": "string",
            "description": "Requis en transport json."
          }
        }
      },
...
    }
  }
}